Run a code formatter, specifically `npx prettier . --write`
Style differs slightly from what I would have preferred. In an ideal
world I would have written a configuration for it at the start of the
project and set pre-commit hooks to automatically run the code formatter
at every commit, but the current guidelines did not easily work for me.

This commit is merely a gesture to symbolize that I do appreciate the
value of enforcing code formatting rules through automated tools in a
project, especially when they be made to work without any headaches. (: